Fixed sending JingleSessionDidStart too early
When we get a description-info stanza we are basically emulating a 180 with
an SDP (sort of). The remote endpoint will now send session-accept with the same
content, which we can ignore, but we wait until we get it to send the DidStart
notification.

This description-info / session-accept workaround if the one implemented by Jitsi,
but it's OK according to the standard. The reason why we need to start the streams
so early is because our RTP engine needs a full SDP we can't just feed ICE candidates
to it, and Jitsi needs the ICE negotiation to be finished *before* sending the
session-accept stanza. Being confronted with this chicken-and-egg problem, this
seemed to be the simplest approach to solve it.