Refactor managing the single account SylkServer currently uses
Create a singleton-ish account instead: DefaultAccount. This account does not register,
subscribe or anything, it's basically a catch-all account acting as a collection of settings.

It's possible that we add more accounts in the future. We might also consider treating accounts
as mere settings containers through a subclass (SettingsAccount), and each room could have one
of these, for example.