Add ability to automatically verify ZRTP endpoints
This setting only applies to the conference application.

In order for automatic verification to take place, the remote SAS must have
been received, using a message of type application/blink-zrtp-sas and it must
match. In addition, there must be a direct path between the endpoint and
SylkServer, that is, no relays.

Last, when this setting is enabled, SylkServer will send a chat message to the
user with the received ZRTP SAS, if no relays are found in the path.